dialer: split dto to separate package
