+++ dialer/filter.go
@@ -2,31 +2,14 @@ package dialer
 
 import (
 	"context"
-	"fmt"
 	"net"
 	"net/http"
+
+	"github.com/SenseUnit/dumbproxy/dialer/dto"
+	"github.com/SenseUnit/dumbproxy/dialer/errors"
 )
 
 type FilterFunc = func(ctx context.Context, req *http.Request, username, network, address string) error
-
-type ErrAccessDenied struct {
-	err error
-}
-
-func (e ErrAccessDenied) Error() string {
-	return fmt.Sprintf("access denied: %v", e.err)
-}
-
-func (e ErrAccessDenied) Unwrap() error {
-	return e.err
-}
-
-type filterContextKey struct{}
-
-type filterContextParams struct {
-	req      *http.Request
-	username string
-}
 
 type FilterDialer struct {
 	f    FilterFunc
@@ -41,9 +24,9 @@ func NewFilterDialer(filterFunc FilterFunc, next Dialer) FilterDialer {
 }
 
 func (f FilterDialer) DialContext(ctx context.Context, network, address string) (net.Conn, error) {
-	req, username := FilterParamsFromContext(ctx)
+	req, username := dto.FilterParamsFromContext(ctx)
 	if ferr := f.f(ctx, req, username, network, address); ferr != nil {
-		return nil, ErrAccessDenied{ferr}
+		return nil, errors.ErrAccessDenied{ferr}
 	}
 	return f.next.DialContext(ctx, network, address)
 }
@@ -54,15 +37,6 @@ func (f FilterDialer) Dial(network, address string) (net.Conn, error) {
 
 func (f FilterDialer) WantsHostname(ctx context.Context, network, address string) bool {
 	return WantsHostname(ctx, network, address, f.next)
-}
-
-func FilterParamsFromContext(ctx context.Context) (*http.Request, string) {
-	params := ctx.Value(filterContextKey{}).(filterContextParams)
-	return params.req, params.username
-}
-
-func FilterParamsToContext(ctx context.Context, req *http.Request, username string) context.Context {
-	return context.WithValue(ctx, filterContextKey{}, filterContextParams{req, username})
 }
 
 var _ Dialer = FilterDialer{}
