+++ dialer/rescache.go
@@ -9,6 +9,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/SenseUnit/dumbproxy/dialer/dto"
 	"github.com/hashicorp/go-multierror"
 	"github.com/jellydator/ttlcache/v3"
 )
@@ -104,7 +105,7 @@ func (nrcd *NameResolveCachingDialer) DialContext(ctx context.Context, network, 
 		return nil, res.err
 	}
 
-	ctx = OrigDstToContext(ctx, address)
+	ctx = dto.OrigDstToContext(ctx, address)
 
 	var dialErr error
 	var conn net.Conn
